+++ hydro/input_hydro_grafic.f90
@@ -48,7 +48,7 @@ subroutine input_hydro_grafic(mdl,r,g,m,ilevel)
   integer::i1,i2,i3,i1_min,i1_max,i2_min,i2_max,i3_min,i3_max
   integer::buf_count
 
-  real(dp)::scale_nH,scale_T2,scale_l,scale_d,scale_t,scale_v
+  real(kind=8)::scale_nH,scale_T2,scale_l,scale_d,scale_t,scale_v
   real(dp)::dx,rr,vx,vy=0,vz=0,ek,ei,pp,xx1,xx2,xx3,dx_loc
 
   real(dp),allocatable,dimension(:,:,:)::init_array
